removed code duplication and fixes
